requirements.txt
aiogram==3.*
python-dotenv
supabase
transformers
torch
torchaudio>=2.2.2
librosa>=0.10.1
soundfile
